cleaning a lot of comments and TODOs
