+++ pybrush/DeapEstimator.py
@@ -4,103 +4,34 @@ sklearn-compatible wrapper for GP analyses.
 See brushgp.cpp for Python (via pybind11) modules that give more fine-grained
 control of the underlying GP objects.
 """
-from sklearn.base import BaseEstimator, ClassifierMixin, RegressorMixin, TransformerMixin
-from sklearn.utils.validation  import check_is_fitted
-# from sklearn.metrics import mean_squared_error
+
+import functools
+
 import numpy as np
 import pandas as pd
-# import deap as dp
+
 from deap import algorithms, base, creator, tools
-# from tqdm import tqdm
+
 from sklearn.metrics import average_precision_score
 from sklearn.preprocessing import MinMaxScaler
-import functools
+
+from sklearn.utils.validation  import check_is_fitted
+from sklearn.base import BaseEstimator, ClassifierMixin, RegressorMixin, \
+                         TransformerMixin
+
+from pybrush.EstimatorInterface import EstimatorInterface
 from pybrush.deap_api import nsga2
 from pybrush import individual
 from pybrush import RegressorEvaluator, ClassifierEvaluator, MultiClassifierEvaluator
 from pybrush import RegressorSelector, ClassifierSelector, MultiClassifierSelector
 from pybrush import RegressorVariator, ClassifierVariator, MultiClassifierVariator
-from pybrush import Parameters, Dataset, SearchSpace
+from pybrush import brush_rng, Parameters, Dataset, SearchSpace
 
-from pybrush import brush_rng
-
-
-class DeapEstimator(BaseEstimator):
+class DeapEstimator(EstimatorInterface, BaseEstimator):
     """
-    This is the base class for Deap-based Brush estimators. 
-    This class shouldn't be called directly; instead, call a child class like 
-    :py:class:`DeapRegressor <brush.estimator.DeapRegressor>` or :py:class:`DeapClassifier <brush.estimator.DeapClassifier>`. 
-    All of the shared parameters are documented here. 
-
-    Parameters
-    ----------
-    mode : str, default 'classification'
-        The mode of the estimator. Used by subclasses
-    pop_size : int, default 100
-        Population size.
-    gens : int, default 100
-        Maximum iterations of the algorithm.
-    verbosity : int, default 0
-        Controls level of printouts.
-    max_depth : int, default 0
-        Maximum depth of GP trees in the GP program. Use 0 for no limit.
-    max_size : int, default 0
-        Maximum number of nodes in a tree. Use 0 for no limit.
-    num_islands : int, default 5
-        Number of independent islands to use in evolutionary framework. 
-        Ignored if `algorithm!="nsga2island"`.
-    mig_prob : float, default 0.05
-        Probability of occuring a migration between two random islands at the
-        end of a generation, must be between 0 and 1.
-    cx_prob : float, default 1/7
-        Probability of applying the crossover variation when generating the offspring,
-        must be between 0 and 1.
-        Given that there are `n` mutations, and either crossover or mutation is 
-        used to generate each individual in the offspring (but not both at the
-        same time), we want to have by default an uniform probability between
-        crossover and every possible mutation. By setting `cx_prob=1/(n+1)`, and
-        `1/n` for each mutation, we can achieve an uniform distribution.
-    mutation_probs : dict, default {"point":1/6, "insert":1/6, "delete":1/6, "subtree":1/6, "toggle_weight_on":1/6, "toggle_weight_off":1/6}
-        A dictionary with keys naming the types of mutation and floating point 
-        values specifying the fraction of total mutations to do with that method.
-        The probability of having a mutation is `(1-cx_prob)` and, in case the mutation
-        is applied, then each mutation option is sampled based on the probabilities
-        defined in `mutation_probs`. The set of probabilities should add up to 1.0.
-    functions: dict[str,float] or list[str], default {}
-        A dictionary with keys naming the function set and values giving the probability
-        of sampling them, or a list of functions which will be weighted uniformly.
-        If empty, all available functions are included in the search space.
-    initialization : {"uniform", "max_size"}, default "uniform" 
-        Distribution of sizes on the initial population. If `max_size`, then every
-        expression is created with `max_size` nodes. If `uniform`, size will be
-        uniformly distributed between 1 and `max_size`.
-    objectives : list[str], default ["error", "size"]
-        list with one or more objectives to use. Options are `"error", "size", "complexity"`.
-        If `"error"` is used, then it will be the mean squared error for regression,
-        and accuracy for classification.
-    algorithm : {"nsga2island", "nsga2", "gaisland", "ga"}, default "nsga2"
-        Which Evolutionary Algorithm framework to use to evolve the population.
-    weights_init : bool, default True
-        Whether the search space should initialize the sampling weights of terminal nodes
-        based on the correlation with the output y. If `False`, then all terminal nodes
-        will have the same probability of 1.0.
-    validation_size : float, default 0.0
-        Percentage of samples to use as a hold-out partition. These samples are used
-        to calculate statistics during evolution, but not used to train the models.
-        The `best_estimator_` will be selected using this partition. If zero, then
-        the same data used for training is used for validation.
-    batch_size : float, default 1.0
-        Percentage of training data to sample every generation. If `1.0`, then
-        all data is used. Very small values can improve execution time, but 
-        also lead to underfit.
-    random_state: int or None, default None
-        If int, then the value is used to seed the c++ random generator; if None,
-        then a seed will be generated using a non-deterministic generator. It is
-        important to notice that, even if the random state is fixed, it is
-        unlikely that running brush using multiple threads will have the same
-        results. This happens because the Operating System's scheduler is
-        responsible to choose which thread will run at any given time, thus 
-        reproductibility is not guaranteed.
+    This is the base class for Brush estimators in python. 
+    
+    Parameters are defined and documented in pybrush.EstimatorInterface.
 
     Attributes
     ----------
@@ -120,50 +51,8 @@ class DeapEstimator(BaseEstimator):
         The toolbox used by DEAP for EA algorithm. 
     """
     
-    def __init__(
-        self, 
-        mode='classification',
-        pop_size=100,
-        gens=100,
-        verbosity=0,
-        max_depth=3,
-        max_size=20,
-        num_islands=1,
-        n_jobs=1,
-        mig_prob=0.05,
-        cx_prob= 1/7,
-        mutation_probs = {"point":1/6, "insert":1/6, "delete":1/6, "subtree":1/6,
-                            "toggle_weight_on":1/6, "toggle_weight_off":1/6},
-        functions: list[str]|dict[str,float] = {},
-        initialization="uniform",
-        algorithm="nsga2",
-        objectives=["error", "size"],
-        random_state=None,
-        weights_init=True,
-        validation_size: float = 0.0,
-        batch_size: float = 1.0
-        ):
-
-        self.pop_size=pop_size
-        self.gens=gens
-        self.verbosity=verbosity
-        self.algorithm=algorithm
-        self.mode=mode
-        self.max_depth=max_depth
-        self.max_size=max_size
-        self.num_islands=num_islands
-        self.mig_prob=mig_prob
-        self.n_jobs=n_jobs
-        self.cx_prob=cx_prob
-        self.mutation_probs=mutation_probs
-        self.functions=functions
-        self.objectives=objectives
-        self.initialization=initialization
-        self.random_state=random_state
-        self.batch_size=batch_size
-        self.weights_init=weights_init
-        self.validation_size=validation_size
-
+    def __init__(self, **kwargs):
+        EstimatorInterface.__init__(self, **kwargs)
 
     def _setup_toolbox(self):
         """Setup the deap toolbox"""
@@ -264,29 +153,13 @@ class DeapEstimator(BaseEstimator):
         # These have a default behavior to return something meaningfull if 
         # no values are set
         self.train_ = self.data_.get_training_data()
-        self.train_.set_batch_size(self.batch_size) # TODO: update batch indexes at the beggining of every generation
+        self.train_.set_batch_size(self.batch_size)
         
         self.validation_ = self.data_.get_validation_data()
 
         self.search_space_ = SearchSpace(self.data_, self.functions_, self.weights_init)
                 
-        self.parameters_ = Parameters()
-        self.parameters_.classification = self.mode == "classification"
-        self.parameters_.n_classes = self.n_classes_
-        self.parameters_.n_jobs = self.n_jobs
-        self.parameters_.pop_size = self.pop_size
-        self.parameters_.gens = self.gens
-        self.parameters_.num_islands = self.num_islands
-        self.parameters_.max_depth = self.max_depth
-        self.parameters_.max_size = self.max_size
-        self.parameters_.objectives = self.objectives
-        self.parameters_.cx_prob = self.cx_prob
-        self.parameters_.mig_prob = self.mig_prob
-        self.parameters_.functions = self.functions
-        self.parameters_.mutation_probs = self.mutation_probs
-
-        if self.random_state is not None:
-            self.parameters_.random_state = self.random_state
+        self.parameters_ = self._wrap_parameters()
 
         if self.mode == "classification":
             self.variator_ = (ClassifierVariator
@@ -308,7 +181,7 @@ class DeapEstimator(BaseEstimator):
 
         # nsga2 and ga differ in the toolbox
         self.archive_, self.logbook_ = nsga2(
-            self.toolbox_, self.gens, self.pop_size, self.cx_prob, 
+            self.toolbox_, self.max_gens, self.pop_size, self.cx_prob, 
             (0.0<self.batch_size<1.0), self.verbosity, brush_rng)
 
         final_ind_idx = 0
