+++ pybrush/_versionstr.py
@@ -1,1 +1,1 @@
-__version__="v1.0"
+__version__="0.1"
