+++ pybrush/deap_api/nsga2.py
@@ -47,52 +47,17 @@ def nsga2(toolbox, NGEN, MU, CXPB, use_batch, verbosity, rnd_flt):
         toolbox.update_current_gen(gen)
 
         # Vary the population
-        # print("--"*20)
-        # print("pop before select")
-        # for p in pop:
-        #     print(p.program.get_model())
-            # print(p.fitness.values)
-            # print(p.fitness.weights)
-            # print(p.fitness.wvalues)
-            # print(p.fitness.rank)
-            # print(p.fitness.loss_v)
-            # print(p.fitness.crowding_dist)
 
         parents = toolbox.select(pop) # , len(pop) # select method from brush's cpp side will use the values in self.parameters_ to decide how many individuals it should select
         
-        # print("--"*20)
-        # print("pop after select")
-        # for p in pop:
-        #     print(p.program.get_model())
-
-        # print("--"*20)
-        # print("selected parents")
-        # for p in parents:
-        #     print(p.program.get_model())
-
         offspring = toolbox.vary_pop(parents)
         offspring = list(toolbox.map(toolbox.assign_fit, offspring))
-
-        # print("--"*20)
-        # print("offspring")
-        # for p in offspring:
-        #     print(p.program.get_model())
 
         # Select the next generation population (no sorting before this step, as 
         # survive==offspring will cut it in half)
         pop = toolbox.survive(pop + offspring)
 
-        # print("--"*20)
-        # print("pop after survival")
-        # for p in pop:
-        #     print(p.program.get_model())
-
         pop = toolbox.migrate(pop)
-
-        # print("--"*20)
-        # print("pop after migration")
-        # for p in pop:
-        #     print(p.program.get_model())
 
         pop.sort(key=lambda x: x.fitness, reverse=True)
