+++ src/bindings/bind_params.cpp
@@ -15,7 +15,7 @@ void bind_params(py::module& m)
         .def(py::init([](){ Brush::Parameters p; return p; }))
         .def_property("verbosity", &Brush::Parameters::get_verbosity, &Brush::Parameters::set_verbosity)
         .def_property("pop_size", &Brush::Parameters::get_pop_size, &Brush::Parameters::set_pop_size)
-        .def_property("gens", &Brush::Parameters::get_gens, &Brush::Parameters::set_gens)
+        .def_property("max_gens", &Brush::Parameters::get_max_gens, &Brush::Parameters::set_max_gens)
         .def_property("max_stall", &Brush::Parameters::get_max_stall, &Brush::Parameters::set_max_stall)
         .def_property("max_time", &Brush::Parameters::get_max_time, &Brush::Parameters::set_max_time)
         .def_property("current_gen", &Brush::Parameters::get_current_gen, &Brush::Parameters::set_current_gen)
