+++ src/ind/individual.cpp
@@ -42,7 +42,7 @@ int Fitness::dominates(const Fitness& b) const
     int flag1 = 0, // to check if this has a better objective
         flag2 = 0; // to check if b    has a better objective
 
-    // TODO: replace comparison of individual values by   using the overloaded  operators (here and in nsga2)
+    // TODO: replace comparison of individual values by using the overloaded  operators (here and in nsga2)
     for (int i=0; i<get_wvalues().size(); ++i) {
         if (get_wvalues().at(i) > b.get_wvalues().at(i)
         || std::isnan(b.get_wvalues().at(i))
