--- src/params.cpp
@@ -1,9 +1,0 @@
-/* Brush
-copyright 2020 William La Cava
-license: GNU/GPL v3
-*/
-#include "params.h"
-
-namespace Brush
-{
-} // Brush
